+++ package.json
@@ -44,6 +44,7 @@
     "browserify": "^10.2.1",
     "gulp": "^3.8.11",
     "gulp-concat": "^2.6.0",
+    "gulp-flatten": "^0.2.0",
     "gulp-jshint": "^1.11.2",
     "gulp-minify-css": "^1.2.1",
     "gulp-nodemon": "^2.0.3",
@@ -53,6 +54,7 @@
     "gulp-shell": "^0.4.1",
     "gulp-uglify": "^1.4.1",
     "gulp-util": "^3.0.6",
+    "run-sequence": "^1.1.2",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0"
   }
