+++ src/jsx/chat_box/chat_box.scss
@@ -1,7 +1,7 @@
 #progress-chatbox-wrapper {
   flex: 2;
   border-style: solid;
-  border-color: palette(Grey, 400);
+  border-color: gray;
   border-width: 0 0 0 1;
   display: flex;
   flex-direction: column;
@@ -20,7 +20,7 @@
 
   & > #message-form-input {
     border-style: solid;
-    border-color: palette(Grey, 400);
+    border-color: gray;
     border-width: 1 0 0 0;
 
     box-sizing: content-box;;
