+++ src/jsx/location_box/location_box.jsx
