+++ src/jsx/location_box/location_box.scss
@@ -3,6 +3,6 @@
   flex: 1;
 
   border-style: solid;
-  border-color: palette(Grey, 400);
+  border-color: gray;
   border-width: 0 0 1 0;
 }
