+++ src/jsx/progress/progress.scss
@@ -3,6 +3,6 @@
   height: 70px;
 
   border-style: solid;
-  border-color: palette(Grey, 400);
+  border-color: gray;
   border-width: 0 0 1 0;
 }
