+++ src/scss/style.scss
@@ -10,8 +10,8 @@
 @import "base/base";
 
 // Components.
-@import "components/locationBox";
-@import "components/progress";
-@import "components/chatBox";
+@import "_components/location_box";
+@import "_components/progress";
+@import "_components/chat_box";
 
 // Pages.
